refactor(index): move alloc_bounds to crate::util so lexical loaders can use it
alloc_bounds's checked_capacity/checked_len bound header-declared sizes
against the true file length before allocating -- useful for any
segment loader parsing unverified on-disk headers, not just vector
segments. Moves it from vector::index (pub(crate)) to util (pub(crate))
and generalizes the "vector segment is corrupted" wording to "segment
is corrupted" so the upcoming DocValuesReader bounds check (#1047) can
reuse it without a vector-specific error message.

No behavior change; existing tests only assert on the "corrupted"
substring.

Refs #1047
