+++ laurus/src/vector/index.rs
@@ -6,7 +6,6 @@
 //! - Vector quantization and compression
 //! - Index optimization and maintenance
 
-pub(crate) mod alloc_bounds;
 pub mod config;
 pub mod factory;
 pub mod field;
